pyproject.toml
[project]
name = "ai-voice-assistant"
version = "0.2.2"
description = "This is an AI voice assistant application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appopener>=1.7",
    "ctranslate2>=4.6.0",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.0.4",
    "faster-whisper>=1.1.1",
    "flet[all]>=0.28.3",
    "gputil>=1.4.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-ollama>=0.3.3",
    "langchain-tavily>=0.2.10",
    "ollama>=0.5.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "pyinstaller>=6.15.0",
    "pywin32>=310",
    "scipy>=1.15.3",
    "setuptools==80.9.0",
    "sounddevice>=0.5.2",
    "speedtest-cli>=2.1.3",
    "torch>=2.7.1",
]

[dependency-groups]
test = [
    "ruff>=0.12.12",
]
